# backend/docs/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ML Evaluation Platform API",
    "description": "API for object detection and segmentation model evaluation platform",
    "version": "1.0.0",
    "contact": {
      "name": "ML Evaluation Platform"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Development server"
    },
    {
      "url": "https://api.ml-eval.cloud",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/api/v1/annotations": {
      "post": {
        "summary": "Create manual annotation",
        "description": "Create user annotation for an image",
        "operationId": "createAnnotation",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAnnotationRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Annotation created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Annotation"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      },
      "get": {
        "summary": "List annotations",
        "description": "Retrieve annotations with filtering options",
        "operationId": "listAnnotations",
        "parameters": [
          {
            "name": "image_id",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "model_id",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "creation_method",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "user",
                "model"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of annotations",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnnotationListResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/deployments": {
      "post": {
        "summary": "Deploy model",
        "description": "Deploy trained model as REST API endpoint",
        "operationId": "deployModel",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeploymentRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Deployment initiated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Deployment"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      },
      "get": {
        "summary": "List deployments",
        "description": "Retrieve list of deployed models",
        "operationId": "listDeployments",
        "responses": {
          "200": {
            "description": "List of deployments",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeploymentListResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/deployments/{deployment_id}": {
      "get": {
        "summary": "Get deployment details",
        "description": "Retrieve deployment status and metrics",
        "operationId": "getDeployment",
        "parameters": [
          {
            "name": "deployment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deployment details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Deployment"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "patch": {
        "summary": "Update deployment",
        "description": "Update deployment configuration or status",
        "operationId": "updateDeployment",
        "parameters": [
          {
            "name": "deployment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeploymentUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Deployment updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Deployment"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/export/annotations": {
      "post": {
        "summary": "Export annotations",
        "description": "Export annotations in standard formats",
        "operationId": "exportAnnotations",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Export file",
            "content": {
              "application/zip": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Annotation": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "image_id": {
            "type": "string",
            "format": "uuid"
          },
          "bounding_boxes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BoundingBox"
            }
          },
          "segments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Segment"
            }
          },
          "class_labels": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "confidence_scores": {
            "type": "array",
            "items": {
              "type": "number",
              "format": "float"
            }
          },
          "creation_method": {
            "type": "string",
            "enum": [
              "user",
              "model"
            ]
          },
          "model_id": {
            "type": "string",
            "format": "uuid"
          },
          "user_tag": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "metadata": {
            "type": "object"
          }
        },
        "required": [
          "id",
          "image_id",
          "class_labels",
          "creation_method",
          "created_at"
        ]
      },
      "BoundingBox": {
        "type": "object",
        "properties": {
          "x": {
            "type": "number",
            "format": "float"
          },
          "y": {
            "type": "number",
            "format": "float"
          },
          "width": {
            "type": "number",
            "format": "float"
          },
          "height": {
            "type": "number",
            "format": "float"
          },
          "class_id": {
            "type": "integer"
          },
          "confidence": {
            "type": "number",
            "format": "float"
          }
        },
        "required": [
          "x",
          "y",
          "width",
          "height"
        ]
      },
      "Segment": {
        "type": "object",
        "properties": {
          "polygon": {
            "type": "array",
            "items": {
              "type": "array",
              "items": {
                "type": "number",
                "format": "float"
              }
            }
          },
          "mask": {
            "type": "string",
            "description": "Base64 encoded mask"
          },
          "class_id": {
            "type": "integer"
          },
          "confidence": {
            "type": "number",
            "format": "float"
          }
        }
      },
      "CreateAnnotationRequest": {
        "type": "object",
        "properties": {
          "image_id": {
            "type": "string",
            "format": "uuid"
          },
          "bounding_boxes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BoundingBox"
            }
          },
          "segments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Segment"
            }
          },
          "class_labels": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "user_tag": {
            "type": "string"
          },
          "metadata": {
            "type": "object"
          }
        },
        "required": [
          "image_id",
          "class_labels"
        ]
      },
...
      "AnnotationListResponse": {
        "type": "object",
        "properties": {
          "annotations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Annotation"
            }
          },
          "total_count": {
            "type": "integer"
          }
        },
        "required": [
          "annotations",
          "total_count"
        ]
      },
...
      "DeploymentRequest": {
        "type": "object",
        "properties": {
          "model_id": {
            "type": "string",
            "format": "uuid"
          },
          "version": {
            "type": "string"
          },
          "configuration": {
            "type": "object",
            "properties": {
              "replicas": {
                "type": "integer",
                "default": 1
              },
              "cpu_limit": {
                "type": "string",
                "default": "1000m"
              },
              "memory_limit": {
                "type": "string",
                "default": "2Gi"
              },
              "gpu_required": {
                "type": "boolean",
                "default": false
              }
            },
            "required": [
              "replicas"
            ]
          },
          "metadata": {
            "type": "object"
          }
        },
        "required": [
          "model_id",
          "version",
          "configuration"
        ]
      },
      "Deployment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "model_id": {
            "type": "string",
            "format": "uuid"
          },
          "endpoint_url": {
            "type": "string"
          },
          "version": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "deploying",
              "active",
              "inactive",
              "failed"
            ]
          },
          "configuration": {
            "type": "object"
          },
          "performance_monitoring": {
            "type": "object"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "metadata": {
            "type": "object"
          }
        },
        "required": [
          "id",
          "model_id",
          "endpoint_url",
          "version",
          "status",
          "configuration",
          "created_at",
          "updated_at"
        ]
      },
      "DeploymentListResponse": {
        "type": "object",
        "properties": {
          "deployments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Deployment"
            }
          },
          "total_count": {
            "type": "integer"
          }
        },
        "required": [
          "deployments",
          "total_count"
        ]
      },
      "DeploymentUpdateRequest": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "active",
              "inactive"
            ]
          },
          "configuration": {
            "type": "object"
          }
        }
      },
      "ExportRequest": {
        "type": "object",
        "properties": {
          "image_ids": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "format": {
            "type": "string",
            "enum": [
              "COCO",
              "YOLO",
              "Pascal VOC"
            ]
          },
          "include_predictions": {
            "type": "boolean",
            "default": false
          },
          "model_id": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "image_ids",
          "format"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": "object"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "error",
          "message",
          "timestamp"
        ]
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "ApiKeyAuth": []
    },
    {
      "BearerAuth": []
    }
  ]
}